# Replaced SoftReference with WeakReference for better performance