--- a/core/src/main/java/fj/Ordering.java
+++ b/core/src/main/java/fj/Ordering.java
@@ -23,6 +23,15 @@ public enum Ordering {
   GT;
 
   public int toInt() { return ordinal() - 1 ; }
+
+  public Ordering reverse() {
+    switch (this) {
+      case LT: return GT;
+      case GT: return LT;
+    }
+    return EQ;
+  }
+
   public static Ordering fromInt(int cmp) {
     return cmp == 0 ? EQ : cmp > 0 ? GT : LT;
   }