--- a/core/src/main/java/fj/data/List.java
+++ b/core/src/main/java/fj/data/List.java
@@ -15,10 +15,8 @@ import fj.P1;
 import fj.P2;
 import fj.Show;
 import fj.Unit;
-import static fj.Function.curry;
-import static fj.Function.constant;
-import static fj.Function.identity;
-import static fj.Function.compose;
+
+import static fj.Function.*;
 import static fj.P.p;
 import static fj.P.p2;
 import static fj.Unit.unit;
@@ -169,9 +167,10 @@ public abstract class List<A> implements Iterable<A> {
    */
   @SuppressWarnings({"unchecked"})
   public final Array<A> toArray() {
-    final Object[] a = new Object[length()];
+    final int length = length();
+    final Object[] a = new Object[length];
     List<A> x = this;
-    for (int i = 0; i < length(); i++) {
+    for (int i = 0; i < length; i++) {
       a[i] = x.head();
       x = x.tail();
     }
@@ -629,14 +628,14 @@ public abstract class List<A> implements Iterable<A> {
   }
 
   /**
-   * Performs a right-fold reduction across this list. This function uses O(length) stack space.
+   * Performs a right-fold reduction across this list.
    *
    * @param f The function to apply on each element of the list.
    * @param b The beginning value to start the application from.
    * @return The final result after the right-fold reduction.
    */
   public final <B> B foldRight(final F<A, F<B, B>> f, final B b) {
-    return isEmpty() ? b : f.f(head()).f(tail().foldRight(f, b));
+    return reverse().foldLeft(flip(f), b);
   }
 
   /**
@@ -1581,7 +1580,9 @@ public abstract class List<A> implements Iterable<A> {
    * @return A list of the given value replicated the given number of times.
    */
   public static <A> List<A> replicate(final int n, final A a) {
-    return n <= 0 ? List.<A>nil() : replicate(n - 1, a).cons(a);
+    List<A> list = List.nil();
+    for (int i = 0; i < n; i++) { list = list.cons(a); }
+    return list;
   }
 
   /**