@jbgi 2015-06-08
but with WeakReference there almost not point to memoize at all (the result can be garaged really fast). Might as well remove the call memo() from the stream implementation.
Another benchmark taking advantage of stream memoization would be useful.