@orionll 2015-06-08
We use a cache of parsed expressions in production (`com.google.common.cache.Cache<String, ASTNode>`). We tested different implementations of cache (WeakReference-based, SoftReference-based) and also no caching at all (each expression is parsed every time). The SoftReference-based cache showed terrible performance, even worse than no caching. The WeakReference-based cache showed the best performance, and we are using it in production now.