@jbgi 2015-06-08
I think there is no general rule, it depends of how expensive it is to re-compute cached values and how frequently you need them. Maybe we need the three memo implementation to let the user chose the most appropriate one:
 - hardMemo() (basically com.google.common.base.Suppliers.MemoizingSupplier) 
 - softMemo() (current implementation) 
 - weakMemo() (your patch) 
It is totally possible that for most use case of streams weakMemo give the best results.