@jbgi 2015-06-08
What I meant is that we should not change the current ```memo()``` behavior (which would be an alias for ```softMemo()```) and maybe instead change Stream implementation to use a new ```weakMemo()``` method on P1.