@orionll 2015-06-08
+1 for adding three distinct methods and removing memo().