@orionll 2015-06-09
So, I think there should be three methods:
* memo() - based on strong references
* weakMemo() - based on weak references
* softMemo() - based on soft references

See my changes.