@jbgi 2015-06-09
@mperry can we change memo() memory behaviour without a full deprecation cycle?