@techtangents 2015-06-09
How about we keep memo(), but add an overload for it that passes in the 
memoization strategy.

e.g.
memo(Memo.hardReference)
memo(Memo.softReference)
memo(Memo.weakReference)

then memo() would be implemented as:

memo(Memo.weakReference)

or whichever behavior you want as default.

> JB Giraudeau <mailto:notifications@github.com>
> 10 June 2015 2:27 am
>
> @mperry <https://github.com/mperry> can we change memo() memory 
> behaviour without a full deprecation cycle?
>
> —
> Reply to this email directly or view it on GitHub 
> <https://github.com/functionaljava/functionaljava/pull/151#issuecomment-110422553>.
>
> Zheka Kozlov <mailto:notifications@github.com>
> 8 June 2015 8:00 pm
>
> Benchmark:
>
> import com.google.common.base.Stopwatch;
> import fj.data.Stream;
>
> public class Main {
> public static void main(String[] args) {
> int to = 3000000;
> Stream<Integer> range = Stream.range(1, 3000000);
>
> Stopwatch sw = Stopwatch.createStarted();
> Integer sum = range.foldLeft(x -> y -> x + y, 0);
> System.out.println("First run: " + sw.stop());
>
> sw.start();
> sum = range.foldLeft(x -> y -> x + y, 0);
> System.out.println("Second run: " + sw.stop());
> }
> }
>
> The result with SoftReference:
>
> |First run: 13,31 s
> Second run: 32,68 s
> |
>
> WeakReference:
>
> |First run: 219,9 ms
> Second run: 388,6 ms
> |
>
> SoftReference is about 70 times slower.
>
> I also performed a profiling session with JVisualVM. With 
> SoftReference there are 125 collections total and garbage collection 
> took 23 seconds! With WeakReference - 29 collections and 25 milliseconds.
>
> ------------------------------------------------------------------------
>
>
>         You can view, comment on, or merge this pull request online at:
>
> https://github.com/functionaljava/functionaljava/pull/151
>
>
>         Commit Summary
>
>   * Replaced SoftReference with WeakReference for performance
>
>
>         File Changes
>
>   * *M* core/src/main/java/fj/P1.java
>     <https://github.com/functionaljava/functionaljava/pull/151/files#diff-0>
>     (6)
>
>
>         Patch Links:
>
>   * https://github.com/functionaljava/functionaljava/pull/151.patch
>   * https://github.com/functionaljava/functionaljava/pull/151.diff
>
> —
> Reply to this email directly or view it on GitHub 
> <https://github.com/functionaljava/functionaljava/pull/151>.
>