@mperry 2015-06-13
I agree with @techtangents, I think we should provide a default memo implementation and then other options to choose from a hard, soft and weak implementation.  Whether this is separate hard, soft and weak methods or a parameter to memo does not matter too much.