@mperry 2015-06-13
@jbgi we can change the implementation of `memo` memory implementation without deprecation.  We don't provide guarantees on the implementation, so changing the implementation when the type signature remains the same I think is fine.