@mperry 2015-06-13
I think this issue is complete enough to merge.  I created #153 to handle hard, soft, weak and default memo implementations.